+++ src/pat/image-crop/index.html
@@ -1,51 +1,62 @@
 <!DOCTYPE html>
 <html lang="en">
-	<head>
-		<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-		<title>Demo page</title>
-        <link rel="stylesheet" href="/style/common.css" type="text/css">
-        <script src="/bundle.js" type="text/javascript" charset="utf-8"></script>
-	</head>
-	<body>
-        <form id="crop-demo-form">
-        	<div class="row">
-            	<div class="seven columns">
-	            	<img src="pool.jpg" class="pat-image-crop" data-pat-image-crop="preview-id: #image-crop-preview;" />
-	            </div>
-	            <div class="five columns pull-one">
-	                <div class="preview">
-	                    <img src="pool.jpg" id="image-crop-preview">
-	                </div>
-	                <div class="row">
-	                	<div class="three columns">
-	                		<fieldset class="vertical condensed">
-		                        <label>x1
-		                            <input type="text" size="2" name="x1" />
-		                        </label>
-		                        <label>y1
-		                            <input type="text" size="2" name="y1" />
-		                        </label>
-		                        <label>x2
-		                            <input type="text" size="2" name="x2" />
-		                        </label>
-	                        </fieldset>
-	                    </div>
-	                    <div class="three columns pull-six">
-	                    	<fieldset class="vertical condensed">
-		                        <label>y2
-		                            <input type="text" size="2" name="y2" />
-		                         </label>
-		                        <label>w
-		                            <input type="text" size="2" name="w" />
-		                        </label>
-		                        <label>h
-		                            <input type="text" size="2" name="h" />
-		                        </label>
-		                    </fieldset>
-	                    </div>
-	                 </div>
-	    	    </div>
-			</div>
-    	</form>
-	</body>
+  <head>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+    <title>Demo page</title>
+    <link rel="stylesheet" href="/style/common.css" type="text/css" />
+    <script src="/bundle-vendors.js" type="text/javascript" charset="utf-8"></script>
+    <script src="/bundle.js" type="text/javascript" charset="utf-8"></script>
+  </head>
+  <body>
+    <form id="crop-demo-form">
+      <div class="row">
+        <div class="seven columns">
+          <img
+            src="pool.jpg"
+            class="pat-image-crop"
+            data-pat-image-crop="preview-id: #image-crop-preview;"
+          />
+        </div>
+        <div class="five columns pull-one">
+          <div class="preview">
+            <img src="pool.jpg" id="image-crop-preview" />
+          </div>
+          <div class="row">
+            <div class="three columns">
+              <fieldset class="vertical condensed">
+                <label
+                  >x1
+                  <input type="text" size="2" name="x1" />
+                </label>
+                <label
+                  >y1
+                  <input type="text" size="2" name="y1" />
+                </label>
+                <label
+                  >x2
+                  <input type="text" size="2" name="x2" />
+                </label>
+              </fieldset>
+            </div>
+            <div class="three columns pull-six">
+              <fieldset class="vertical condensed">
+                <label
+                  >y2
+                  <input type="text" size="2" name="y2" />
+                </label>
+                <label
+                  >w
+                  <input type="text" size="2" name="w" />
+                </label>
+                <label
+                  >h
+                  <input type="text" size="2" name="h" />
+                </label>
+              </fieldset>
+            </div>
+          </div>
+        </div>
+      </div>
+    </form>
+  </body>
 </html>
